fix: 1.0.0 M2 sections never drew, and nested sidebar sub-tabs were unreachable
M2 era-100 section layout was guessed, not recovered ("Per-field layout not
fully recovered from Ghidra; using standard M2SkinSection fields"), and the
guess was wrong. M2SkinSection stores every field as uint16: indexStart @0x08
and indexCount @0x0A. The reader took uint32 reads at 0x08/0x0C, folding
indexCount into indexStart's high bits — GnomeFemale.m2 resolved an indexStart
of ~393216 against a far smaller index buffer, so the runtime bridge's bounds
check rejected every section and the load died with "contained no drawable
embedded sections". Level, which carries the high bits of vertexStart and
indexStart, was ignored entirely; it is now applied.

Added M2Era100ModelReaderTests — the era-100 fixture tests tasks T003/T004 ask
for and that never existed. The 9 passing M2Era1121 tests only ever covered
3X/MDLX/2X/4X, so nothing exercised the 1.0.0 path. 3 of the 4 new tests fail
against the old layout.

Era-100 parse failure was documented as terminal but wasn't: the throw sat
inside an outer try whose catch(Exception) swallowed it and logged "falling
through to legacy parser", handing 1.0.0 bytes to a different M2/MDX layout —
exactly what the contract forbids. Detection and parse are now separated so the
terminal exception escapes. The MDX-adaptation path also clamped index fields
to ushort.MaxValue when its targets were already int.

Sidebar: Tools > Terrain and Tools > Utilities are a second nesting level but
re-read the parent's _activeBottomTabIndex, so selecting "Utilities" (index 4)
rendered UtilitiesBottomTab 4 (Taxi) and "Terrain" (index 3) rendered
TerrainBottomTab 3 (Weak Signal). Capture Automation, Asset Catalog, Runtime
Stats, Minimap, Log, Perf, Render Quality and the MCNK tab were unreachable —
GetTerrainBottomTabLabels() existed but was never called, because the
second-level tab strip was never drawn. Each nested level now owns its index
and draws its own strip.

Solution builds 0 errors. The 11 Core test failures are pre-existing, confirmed
by running them against the parent commit.

Unproven: no 1.0.0 client is staged locally, so the reader fix is verified
against synthetic fixtures and the documented layout, not a real GnomeFemale.m2.

+++ wow-viewer/src/core/WowViewer.Core/M2/M2Era100Geometry.cs
@@ -71,11 +71,13 @@ public readonly record struct M2Era100Vertex(
 /// <summary>
 /// Render section / geometry group (0x20 bytes on disk). Defines a contiguous slice
 /// of the vertex/index buffers that shares a material.
+/// VertexStart and IndexStart are stored on disk as uint16 with their high bits in Level;
+/// both are already combined here.
 /// </summary>
 public readonly record struct M2Era100Section(
     ushort SubmeshId,
     ushort Level,
-    ushort VertexStart,
+    uint VertexStart,
     ushort VertexCount,
     uint IndexStart,
     uint IndexCount);
